chore: refresh 버튼을 textField 안으로
